Add sendResult to AceComm callback forwarding
Most people probably interface with AceComm rather than CTL directly, so
it makes sense to pass the send result from CTL fully through the chain.

One downside is that because AceComm currently appends its own textLen
argument, we're forced to effectively sandwich it by appending
sendResult to the end of the list.