Move Link method off Ring metatable
There's a bug in the upgrade logic where existing rings created by older
versions of the library never have their metatables upgraded. As such, a
load of a version older than v25 followed by v25 or newer will result in
a call to a nil 'Link' method on rings.

Rather than think through the ramifications of just setmetatable-ing
rings again unilaterally on upgrade (though that is probably a perfectly
sane idea and should work), for now let's just make the new Link method
a local function instead as that's the safer option.