Improve handling of errored send calls
With the changes to implement prefix throttle support the logic around
despooling was changed such that messages in a queue are peeked, sent,
and then popped based upon whether or not it was throttled.

The rationale behind that approach was simply performance; currently
message pipes are an array of messages where a pop down-shifts every
element in the queue. If we had to re-queue a message that would result
in re-shifting everything back up unnecessarily.

This change however makes things a bit worse in the case of send calls
that encounter an error in the C API itself - eg. due to invalid
parameters.

In such a case, because we've not popped the message prior to sending it
we end up in a situation where the message sticks in the queue, and CTL
now infinitely tries to re-send it expecting a different result.

To alleviate this concern, the PerformSend and MapToSendResult function
have been altered slightly to now xpcall the send function and check
whether or not it actually succeeded.

- In the event where a message failed to send due to an error, the
  global error handler is invoked and CTL will continue its own logic
  with the assumption that the message generally failed to send. It will
  be dequeued and not retried.

- In the event where no error occured, the logic is ultimately the same
  as before, just now centralized in the MapToSendResult function.